backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.1
httpx==0.27.0
redis==5.0.4
qdrant-client==1.9.1
anthropic==0.28.0
langgraph==0.1.5
langchain-core==0.2.5
langchain-anthropic==0.1.15
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.4
reportlab==4.2.0
PyPDF2==3.0.1
pdfminer.six==20231228
matplotlib==3.9.0
Pillow==10.3.0
aiofiles==23.2.1
tenacity==8.3.0
structlog==24.2.0
